% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framer.R
\name{lift_data}
\alias{lift_data}
\title{Lift up}
\usage{
lift_data(data, where, from = NULL, to = NULL)
}
\arguments{
\item{data}{A tidy data frame.}

\item{where}{The tag where a value is located; eg, 'author'.}

\item{from}{The tag to extract a value from. This will usually be 'fulltext'.}

\item{to}{The new field to put the value into.}
}
\value{

}
\description{
Lift up
}
