% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framer.R
\name{summarize_llr}
\alias{summarize_llr}
\title{Summarize the log-likelihood ratio across a grouped data frame}
\usage{
summarize_llr(data, token, count = rep(1, n()))
}
\arguments{
\item{data}{A data frame}

\item{token}{The column indicating a token.}

\item{count}{The column indicating wordcount data.}
}
\value{
A dataframe with the supplied grouping and a log-likelihood for each token in that grouping.
Strongly positive numbers are over-represented; strongly negative numbers are under-represented.
}
\description{
Summarize the log-likelihood ratio across a grouped data frame
}
