% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/framer.R
\name{TEIdy}
\alias{TEIdy}
\title{Load an XML document as a data.frame}
\usage{
TEIdy(fname, ignore = c(), discard = c(), keep = c())
}
\arguments{
\item{fname}{A filename}

\item{ignore}{Tags that, when encountered, will not be added as columns to the tidy representation.}

\item{discard}{Tags that, when encountered, will be discarded from the tidy representation. discard='teiHeader', for
example, will prevent header information from being included.}

\item{keep}{Tags to *keep* in the tidy representation. If this is used, 'ignore' and 'discard' arguments will apply
only inside tags defined by keep.}
}
\value{
A tibble, with a 'text' column indicating the lowest of text address found in the document.
}
\description{
Load an XML document as a data.frame
}
\examples{

tomorrow = system.file("extdata", "Tomorrow_and_tomorrow.xml", package="TEIdy")
TEIdy(tomorrow)

}
