% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEIdytext.R
\docType{package}
\name{TEIdytext}
\alias{TEIdytext}
\alias{TEIdytext-package}
\title{TEIdytext: Work with XML-tagged documents in the tidytext framework.}
\description{
TEIdytext: Work with XML-tagged documents in the tidytext framework.
}
